+++ Assets/Python/components/Provinces.py
@@ -1,6 +1,5 @@
-from Core import civilization, civilizations, player, year, cities
+from Core import get_scenario, civilization, civilizations, player, year, cities
 from RFCUtils import refreshStabilityOverlay
-from Scenario import get_scenario
 from ProvinceMapData import PROVINCES_MAP
 from CoreTypes import Province, Event, Scenario, ProvinceType
