+++ Assets/Python/components/Stability.py
@@ -1,7 +1,7 @@
 # Rhye's and Fall of Civilization: Europe - Stability
 
 from CvPythonExtensions import *
-from Core import civilization, civilizations, message, human, cities, text
+from Core import get_scenario, civilization, civilizations, message, human, cities, text
 
 from Consts import MessageData
 from CoreTypes import (
@@ -25,7 +25,6 @@ from ProvinceMapData import PROVINCES_MAP
 from RFCUtils import collapseImmune, getLastRespawnTurn, getUniqueBuilding, killAndFragmentCiv
 import RiseAndFall
 import Provinces
-from Scenario import get_scenario
 from Secession import revoltCity
 
 rnf = RiseAndFall.RiseAndFall()
